Fix grainsForSlots int*int overflow in slot count computation
numSlots * slotSize can exceed INT_MAX for realistic production
values (e.g. 10K slots * 500KB = 5B). Cast to int64_t before
multiplication to match Python's arbitrary-precision arithmetic.

+++ cpp/tensorrt_llm/batch_manager/kv_cache_manager_v2/storage/core.cpp
@@ -698,7 +698,7 @@ int64_t CacheLevelStorage::grainsForSlots(int numSlots, std::vector<int> const& 
 {
     int64_t total = 0;
     for (auto s : slotSizeList)
-        total += divUp(numSlots * s, granularity);
+        total += divUp(static_cast<int64_t>(numSlots) * s, static_cast<int64_t>(granularity));
     return total;
 }
